wallet/requirements.txt
Django==5.1.6
django-filter==24.3
django-split-settings==1.3.2
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
gunicorn==20.1.0
flake8==6.0.0